feat: load nbt_b64 from recipes.json
