+++ Allay-Server/src/main/java/org/allaymc/server/item/recipe/AllayRecipeRegistry.java
@@ -16,7 +16,9 @@ import org.allaymc.api.item.descriptor.ItemTagDescriptor;
 import org.allaymc.api.item.init.SimpleItemStackInitInfo;
 import org.allaymc.api.item.recipe.*;
 import org.allaymc.api.item.registry.ItemTypeRegistry;
+import org.allaymc.api.utils.NbtUtils;
 import org.allaymc.server.item.type.AllayItemType;
+import org.cloudburstmc.nbt.NbtMap;
 import org.cloudburstmc.protocol.bedrock.data.inventory.crafting.CraftingDataType;
 import org.cloudburstmc.protocol.bedrock.data.inventory.crafting.recipe.RecipeData;
 import org.cloudburstmc.protocol.bedrock.packet.CraftingDataPacket;
@@ -167,12 +169,13 @@ public class AllayRecipeRegistry implements RecipeRegistry {
         Objects.requireNonNull(itemType, "Unknown item type: " + itemId);
         var count = jsonObject.get("count").getAsInt();
         var meta = jsonObject.has("damage") ? jsonObject.get("damage").getAsInt() : 0;
-        // TODO: nbt_b64
+        var nbtMap = jsonObject.has("nbt_b64") ? NbtUtils.base64ToNbt(jsonObject.get("nbt_b64").getAsString()) : NbtMap.EMPTY;
         return itemType.createItemStack(
                 SimpleItemStackInitInfo
                         .builder()
                         .count(count)
                         .meta(meta)
+                        .extraTag(nbtMap)
                         .build()
         );
     }
